requirements.txt
streamlit
youtube-transcript-api
anthropic
python-dotenv